Wait 2 minutes for redis to be up and running
Why? because local testing is working stably without that 1 minute wait, meaning the http service is immediately ready for us to send post request to. Why redis doesn't contain expected data in github workflow? Probably because by the time we post, redis is not up and running hence can't receive the data. If I give redis 1 mins to be up before posting, would it receive the data and pass data verification?

Signed-off-by: Keran Yang <yangkr920208@gmail.com>