Switch to use old log verification
In this change, we don't wait 1 min. We directly send msgs to http service. We use log to verify output. If this test is stable. Then we prove http sending requests is stable and the problem comes from redis(if not wait 1 minute, then redis somehow can't receive the data).

Signed-off-by: Keran Yang <yangkr920208@gmail.com>